refactor: move useAppContext hook
