+++ src/client/features/Album/template.tsx
@@ -13,7 +13,7 @@ import { useTranslations } from 'client/hooks/useTranslations';
 import { HorizontalCard } from 'client/components/molecules/cards/HorizontalCard';
 import { useIntl } from 'client/hooks/useIntl';
 import { GridImage } from 'client/components/atoms/images/GirdImage';
-import { useAppContext } from 'client/hooks/useAppContext';
+import { useAppContext } from 'client/store/app/useAppContext';
 import { PageHelmet } from 'client/layouts/PageHelmet';
 
 export const AlbumPage: React.FC<AlbumPageProps> = props => {
