+++ src/client/features/Search/template.tsx
@@ -9,7 +9,7 @@ import { Typography } from 'client/components/atoms/Typography';
 import { PageContent } from 'client/components/templates/Page';
 import { commonStyles, useAppTheme } from 'client/styles/tokens';
 import { useTranslations } from 'client/hooks/useTranslations';
-import { useAppContext } from 'client/hooks/useAppContext';
+import { useAppContext } from 'client/store/app/useAppContext';
 import { PageHelmet } from 'client/layouts/PageHelmet';
 
 export type SearchProps = {
